+++ api/src/integration/java/org/pmiops/workbench/NotebooksIntegrationTest.java
@@ -7,9 +7,9 @@ import org.pmiops.workbench.db.dao.WorkspaceDao;
 import org.pmiops.workbench.firecloud.FireCloudService;
 import org.pmiops.workbench.firecloud.FirecloudApiClientFactory;
 import org.pmiops.workbench.google.StorageConfig;
-import org.pmiops.workbench.legacy_leonardo_client.LeonardoApiClient;
-import org.pmiops.workbench.legacy_leonardo_client.LeonardoApiClientFactory;
-import org.pmiops.workbench.legacy_leonardo_client.LeonardoApiClientImpl;
+import org.pmiops.workbench.leonardo.LeonardoApiClient;
+import org.pmiops.workbench.leonardo.LeonardoApiClientFactory;
+import org.pmiops.workbench.leonardo.LeonardoApiClientImpl;
 import org.pmiops.workbench.utils.mappers.LeonardoMapper;
 import org.pmiops.workbench.workspaces.WorkspaceAuthService;
 import org.pmiops.workbench.workspaces.resources.UserRecentResourceService;
