+++ api/src/main/java/org/pmiops/workbench/initialcredits/InitialCreditsExpirationServiceImpl.java
@@ -15,7 +15,7 @@ import org.pmiops.workbench.db.model.DbUserInitialCreditsExpiration.Notification
 import org.pmiops.workbench.db.model.DbWorkspace;
 import org.pmiops.workbench.exceptions.WorkbenchException;
 import org.pmiops.workbench.institution.InstitutionService;
-import org.pmiops.workbench.legacy_leonardo_client.LeonardoApiClient;
+import org.pmiops.workbench.leonardo.LeonardoApiClient;
 import org.pmiops.workbench.mail.MailService;
 import org.pmiops.workbench.model.BillingStatus;
 import org.pmiops.workbench.utils.BillingUtils;
