+++ api/src/main/java/org/pmiops/workbench/leonardo/LeonardoApiClient.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.legacy_leonardo_client;
+package org.pmiops.workbench.leonardo;
 
 import java.util.List;
 import java.util.Map;
