+++ api/src/main/java/org/pmiops/workbench/leonardo/LeonardoApiClientFactory.java
@@ -1,10 +1,11 @@
-package org.pmiops.workbench.legacy_leonardo_client;
+package org.pmiops.workbench.leonardo;
 
 import com.google.auth.oauth2.OAuth2Credentials;
 import jakarta.inject.Provider;
 import java.io.IOException;
 import org.pmiops.workbench.config.WorkbenchConfig;
 import org.pmiops.workbench.firecloud.FirecloudApiClientFactory;
+import org.pmiops.workbench.legacy_leonardo_client.ApiClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
