+++ api/src/main/java/org/pmiops/workbench/leonardo/LeonardoAppUtils.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.legacy_leonardo_client;
+package org.pmiops.workbench.leonardo;
 
 import java.util.Map;
 import java.util.Optional;
