+++ api/src/main/java/org/pmiops/workbench/leonardo/LeonardoConfig.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.legacy_leonardo_client;
+package org.pmiops.workbench.leonardo;
 
 import com.google.common.collect.ImmutableList;
 import jakarta.servlet.http.HttpServletRequest;
