+++ api/src/main/java/org/pmiops/workbench/leonardo/LeonardoLabelHelper.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.legacy_leonardo_client;
+package org.pmiops.workbench.leonardo;
 
 import jakarta.annotation.Nullable;
 import java.util.HashMap;
