+++ api/src/test/java/org/pmiops/workbench/api/AppsControllerTest.java
@@ -16,8 +16,8 @@ import org.pmiops.workbench.db.model.DbWorkspace;
 import org.pmiops.workbench.exceptions.FailedPreconditionException;
 import org.pmiops.workbench.exceptions.ForbiddenException;
 import org.pmiops.workbench.interactiveanalysis.InteractiveAnalysisService;
-import org.pmiops.workbench.legacy_leonardo_client.LeonardoApiClient;
-import org.pmiops.workbench.legacy_leonardo_client.LeonardoApiHelper;
+import org.pmiops.workbench.leonardo.LeonardoApiClient;
+import org.pmiops.workbench.leonardo.LeonardoApiHelper;
 import org.pmiops.workbench.model.AppType;
 import org.pmiops.workbench.model.CreateAppRequest;
 import org.pmiops.workbench.workspaces.WorkspaceAuthService;
