+++ api/src/test/java/org/pmiops/workbench/leonardo/LeonardoAppUtilsTest.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.legacy_leonardo_client;
+package org.pmiops.workbench.leonardo;
 
 import static com.google.common.truth.Truth.assertThat;
