+++ api/tools/src/main/java/org/pmiops/workbench/tools/ListDisks.java
@@ -23,12 +23,12 @@ import org.pmiops.workbench.db.model.DbWorkspace;
 import org.pmiops.workbench.firecloud.FireCloudServiceImpl;
 import org.pmiops.workbench.firecloud.FirecloudApiClientFactory;
 import org.pmiops.workbench.google.GoogleConfig;
-import org.pmiops.workbench.legacy_leonardo_client.LeonardoApiClient;
-import org.pmiops.workbench.legacy_leonardo_client.LeonardoApiClientFactory;
-import org.pmiops.workbench.legacy_leonardo_client.LeonardoApiClientImpl;
-import org.pmiops.workbench.legacy_leonardo_client.LeonardoConfig;
-import org.pmiops.workbench.legacy_leonardo_client.LeonardoLabelHelper;
 import org.pmiops.workbench.legacy_leonardo_client.model.LeonardoListPersistentDiskResponse;
+import org.pmiops.workbench.leonardo.LeonardoApiClient;
+import org.pmiops.workbench.leonardo.LeonardoApiClientFactory;
+import org.pmiops.workbench.leonardo.LeonardoApiClientImpl;
+import org.pmiops.workbench.leonardo.LeonardoConfig;
+import org.pmiops.workbench.leonardo.LeonardoLabelHelper;
 import org.pmiops.workbench.sam.SamRetryHandler;
 import org.pmiops.workbench.utils.mappers.LeonardoMapperImpl;
 import org.springframework.boot.CommandLineRunner;
